Update kill-all-connections.sql
launch the conditional query if the table exists.
this will solve the script in ./bin/bootstrap if launch for the first time a the table does not exist.

+++ backend-flask/db/kill-all-connections.sql
@@ -1,7 +1,14 @@
-SELECT pg_terminate_backend(pid) 
-FROM pg_stat_activity 
-WHERE 
--- don't kill my own connection!
-pid <> pg_backend_pid()
--- don't kill the connections to other databases
-AND datname = 'cruddur';
+
+
+DO $$ 
+BEGIN
+    IF EXISTS (SELECT 1 FROM pg_database WHERE datname = 'cruddur') THEN
+          SELECT pg_terminate_backend(pid) 
+          FROM pg_stat_activity 
+          WHERE 
+          -- don't kill my own connection!
+          pid <> pg_backend_pid()
+          -- don't kill the connections to other databases
+          AND datname = 'cruddur';
+    END IF;
+END $$;
